Revert "Feature/tests"
