+++ README.md
@@ -1,121 +1,1 @@
-![](https://img.shields.io/badge/Microverse-blueviolet)
-
-# MicroFlix
-
-> This is a demo project illustrating a movie app. The app list out the movies that one can comment on it and like it. It loads diffrent movies eaach time the user reloads the page.
-
-## Built With
-
-- Major languages HTML, CSS and Javascript
-- Frameworks JEST
-- Technologies used Javascript webpack
-
-## Live Demo (if available)
-
-[Live Demo Link](https://bhskr44.github.io/capstone-project-two/dist)
-
-## Getting Started
-
-**This is an example of how you may give instructions on setting up your project locally.**
-**Modify this file to match your project, remove sections that don't apply. For example: delete the testing section if the currect project doesn't require testing.**
-
-To get a local copy up and running follow these simple example steps.
-
-### Prerequisites
-
-- You need a GitHub account.
-- You need to have a Git client installed on your computer.
-- You need to have a copy of the HTML CSS website codebase on your computer.
-- A code Editor of your choice.
-
-### Setup
-
-```
-git clone https://github.com/bhskr44/capstone-project-two.git
-cd capstone-project-two
-```
-
-### Install
-
-- Install all the dependencys required for the website. BY using the command below.
-  ```
-  npm install
-  ```
-- Open the index.html file in a browser.
-
-### Usage
-
-For making any changes to the project the clone the project-
-
-- Folder Structure
-
-  ```
-  __tests__
-
-  dist
-
-  src
-    css
-     style.css  //Responsible for all the stylings
-
-    module
-      __mocks__
-          comments.js //jest mocks
-          loadApijs   //jest mocks
-
-      checkForLikes.js
-      comments.js
-      getLikes.js
-      getMovieDetails.js
-      liked.js
-      loadApijs
-      postComment.js
-      postCommentMethod.js
-      randomString.js
-
-    index.html    //main HTML File
-    index.js      //main js File
-
-  webpack.config.js
-
-  package.json
-
-  babel.config.js
-  ```
-
-### Run tests
-
-If you want to work with Chrome Developer Tools, simply run the HTML document in Google Chrome and right-click the HTML element you want to inspect. Click on "Inspect" and you will have the tools to run, analyze, and even debug the code.
-
-### Deployment
-
-If you like to deploy the website to the github pages please follow the link [here](https://docs.github.com/en/pages/getting-started-with-github-pages/configuring-a-publishing-source-for-your-github-pages-site).
-
-## Authors
-
-👤 **Bhaskar Ranjan Bora**
-
-- Website: [https://bhskr44.github.io/](https://bhskr44.github.io/)
-- GitHub: [@bhskr44](https://github.com/bhskr44)
-- Twitter: [@bhskr44](https://twitter.com/bhskr44)
-- LinkedIn: [bhskr44](https://linkedin.com/in/bhskr44)
-
-👤 **Robert Alabi**
-
-- GitHub: [@githubhandle](https://github.com/Alabi12)
-- Twitter: [@twitterhandle](https://twitter.com/wolo_robert)
-- LinkedIn: [LinkedIn](https://linkedin.com/in/robert-alabi)
-
-## 🤝 Contributing
-
-Contributions, issues, and feature requests are welcome!
-
-## Show your support
-
-Give a ⭐️ if you like this project!
-
-## Acknowledgments
-
-- Microverse for proving this platform.
-
-## 📝 License
+# capstone-project-two
