+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "private": true,
   "scripts": {
-    "test": "jest",
+    "test": "echo \"Error: no test specified\" && exit 1",
     "build": "webpack",
     "start": "webpack serve --open"
   },
@@ -12,17 +12,13 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@babel/core": "^7.20.2",
     "@babel/plugin-transform-modules-commonjs": "^7.19.6",
-    "@babel/preset-env": "^7.20.2",
-    "@jest/globals": "^29.3.1",
     "babel-eslint": "^10.1.0",
-    "babel-jest": "^29.3.1",
     "css-loader": "^6.7.1",
     "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.26.0",
-    "hint": "^2.0.0",
+    "hint": "^7.1.3",
     "html-webpack-plugin": "^5.5.0",
     "jest": "^29.3.1",
     "style-loader": "^3.3.1",
@@ -30,19 +26,12 @@
     "stylelint-config-standard": "^21.0.0",
     "stylelint-csstree-validator": "^1.9.0",
     "stylelint-scss": "^3.21.0",
-    "ts-jest": "^29.0.3",
     "webpack": "^5.75.0",
     "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^4.11.1"
   },
   "dependencies": {
-    "cross-fetch": "^3.1.5",
     "lodash": "^4.17.21",
-    "mini-css-extract-plugin": "^2.6.1",
-    "node-fetch": "^3.3.0",
-    "sweetalert2": "^11.4.8"
-  },
-  "devServer": {
-    "host": "127.0.0.1"
+    "mini-css-extract-plugin": "^2.6.1"
   }
 }
