+++ src/css/style.css
@@ -235,7 +235,7 @@ h3 {
   width: 100%;
   height: 100%;
   border-radius: 5px;
-  object-fit: contain;
+  object-fit: cover;
 }
 
 .popup-details {
@@ -296,18 +296,4 @@ textarea {
 
 .hide {
   display: none;
-}
-
-.fade-in-image {
-  animation: fadeIn 5s;
-}
-
-@keyframes fadeIn {
-  0% {
-    opacity: 0;
-  }
-
-  100% {
-    opacity: 1;
-  }
 }
