--- src/module/checkForLikes.js
@@ -1,13 +1,0 @@
-const checkForLikes = (movieId) => {
-  const likes = JSON.parse(localStorage.getItem('likes'));
-  let like = 0;
-  likes.forEach((element) => {
-    if (element.item_id === movieId) {
-      like = element.likes;
-      // like += 1;
-    }
-  });
-  return like;
-};
-
-export default checkForLikes;
