--- src/module/randomString.js
@@ -1,12 +1,0 @@
-const randomString = () => 'p';
-// //Code to be Used in the Future
-// const randomString = (length) => {
-//   let result = '';
-//   const characters = 'abcdefghijklmnopqrstuvwxyz';
-//   const charactersLength = characters.length;
-//   for (let i = 0; i < length; i += 1) {
-//     result += characters.charAt(Math.floor(Math.random() * charactersLength));
-//   }
-//   return result;
-// };
-export default randomString;
